Починка безгильзовых патронов (#531)
* Fix .308 caliber ammo

Исправление путаницы с патронами .308 и невозможностью зарядить их в мориту.

* Update rifle.dm - 

убираем дубли


* Возвращение HorizonX

Возвращение офицерского револьвера Интекам.

* Починка безгильзовых патронов

Теперь безгильзовые патроны, если не произведен выстрел , будут извлекатся из ствола а не пропадать.

---------

+++ code/modules/projectiles/ammunition/caseless/_caseless.dm
@@ -4,7 +4,14 @@
 	heavy_metal = FALSE
 
 /obj/item/ammo_casing/caseless/on_eject()
-	qdel(src)
+	// [CELADON-EDIT] - CELADON_FIXES
+	// qdel(src) // CELADON-EDIT - ORIGINAL
+	if(BB)  // Проверяем, что гильза не пустая
+		forceMove(drop_location()) // Если гильза не spent, выбрасываем ее на землю.
+		bounce_away(TRUE)
+	else
+		qdel(src) // Если гильза spent, удаляем ее.
+		// [/CELADON-EDIT]
 
 // Overridden; caseless ammo does not distinguish between "live" and "empty"/"spent" icon states (because it has no casing).
 /obj/item/ammo_casing/caseless/update_icon_state()
